move vcr requirement into own module
